Better `cd` in installer
With `cd "$(dirname $)"/..` it works bad, because `$(dirname $)` will return `.` so it implies that to correctly works, you have to be in the `script` directory, otherwise if you're e.g. in `$HOME/.dotfiles` you'll go in `$HOME` and then look for all `install.sh`, which may not be what you intended to do.

Using `$(dirname $0)` instead ensure that you'll go one level higher than the `install` script.